Added ChunkMap serde, factored out ChunkMap into separate file
