+++ Cargo.toml
@@ -23,8 +23,11 @@ serde_derive = "^1.0"
 toml = "^0.4"
 bincode = "^1.1"
 net2 = "^0.2"
-hashbrown = "^0.3.0"
 ncollide3d = "^0.19"
 nphysics3d = "^0.11"
 num = "^0.2"
 derive_more = "^0.14"
+
+[dependencies.hashbrown]
+version = "^0.3.0"
+features = ["serde"]
