+++ src/sim/chunk/mod.rs
@@ -2,9 +2,6 @@ use crate::block::{Block, BlockId, BlockRegistry};
 use crate::CHUNK_SIZE;
 use crate::Vertex;
 use serde_derive::{Deserialize, Serialize};
-use hashbrown::hash_map::HashMap;
-use hashbrown::hash_map::Entry;
-use hashbrown::hash_map::DefaultHashBuilder;
 
 pub type BlockData = BlockId;
 pub type ChunkFragment = [BlockData; CHUNK_SIZE];
@@ -14,50 +11,7 @@ pub type ChunkSidesArray = [[[u8; CHUNK_SIZE]; CHUNK_SIZE]; CHUNK_SIZE];
 mod pos;
 pub use pos::*;
 
-pub struct ChunkMap{
-    map : HashMap<ChunkPos, ChunkState>
-}
-
-impl ChunkMap {
-
-    pub fn new() -> ChunkMap {
-        ChunkMap {
-            map : HashMap::default()
-        }
-    }
-
-    pub fn get_mut(&mut self, pos : &ChunkPos) -> Option<&mut ChunkState> {
-        self.map.get_mut(pos)
-    }
-
-    pub fn get(&self, pos : &ChunkPos) -> Option<&ChunkState> {
-        self.map.get(pos)
-    }
-
-    pub fn entry(&mut self, pos : ChunkPos) -> Entry<ChunkPos, ChunkState, DefaultHashBuilder> {
-        self.map.entry(pos)
-    }
-
-    pub fn retain<F : FnMut(&ChunkPos, &mut ChunkState) -> bool>(&mut self, f : F) {
-        self.map.retain(f)
-    }
-
-    pub fn contains_key(&self, pos : &ChunkPos) -> bool {
-        return self.map.contains_key(pos);
-    }
-
-    pub fn set(&mut self, pos : ChunkPos, i_pos : InnerChunkPos, block : BlockId) {
-        match self.get_mut(&pos) {
-            None => {print!("Failed to set {:?} : {:?} to {:?}!\n", pos, i_pos, block);},
-            Some(ref mut state) => {
-                print!("Setting {:?} : {:?} to {:?}!\n", pos, i_pos, block);
-                state.set(block, i_pos);
-            }
-        }
-
-    }
-
-}
+pub mod map;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub enum ChunkState {
